backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
google-generativeai>=0.3.0
pydantic>=2.0.0
gunicorn>=21.2.0
